Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0

# Google Cloud
google-cloud-aiplatform==1.51.0
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3
google-cloud-firestore==2.16.0
firebase-admin==6.5.0

# Performance
uvloop==0.19.0
httptools==0.6.1
pillow==10.3.0

# Rate limiting
slowapi==0.1.9